feat: migrate controller to TS
